requirements.txt
bidict==0.23.1
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
dnspython==2.7.0
eventlet==0.37.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.2
python-dotenv==1.2.1
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
Werkzeug==3.0.4
wsproto==1.2.0
zope.event==5.0
zope.interface==7.1.1
